Create RegionalAd.tsx, splitting Ad.tsx into two files
