+++ dotcom-rendering/src/amp/components/Ad.tsx
@@ -1,7 +1,4 @@
-import { ClassNames } from '@emotion/react';
-
 import { adJson, stringify } from '@root/src/amp/lib/ad-json';
-import { regionClasses } from '@root/src/amp/lib/region-classes';
 
 // Largest size first
 const inlineSizes = [
@@ -12,11 +9,6 @@ const inlineSizes = [
 
 // Note: amp-sticky-ad has max height of 100
 const stickySizes = [{ width: 320, height: 50 }]; // Mobile Leaderboard
-
-type AdRegion = 'US' | 'AU' | 'ROW';
-
-// Array of possible ad regions
-const adRegions: AdRegion[] = ['US', 'AU', 'ROW'];
 
 const dfpAdUnitRoot = 'theguardian.com';
 
@@ -34,28 +26,6 @@ const preBidServerPrefix = 'https://prebid.adnxs.com/pbs/v1/openrtb2/amp';
 const permutiveURL = 'https://guardian.amp.permutive.com/rtc?type=doubleclick';
 const amazonConfig = {
 	aps: { PUB_ID: '3722', PARAMS: { amp: '1' } },
-};
-
-/**
- * Determine the Placement ID that is used to look up a given stored bid request
- *
- * Stored bid requests are stored by the prebid server instance and each is
- * keyed by a placement ID. This placement ID corresponds to the tag id parameter
- * provided on the client
- *
- * @param adRegion The advertising region - different regions are covered by different
- * stored bid requests
- * @returns The placement id for an ad, depending on its ad region
- */
-const getPlacementIdByAdRegion = (adRegion: AdRegion): number => {
-	switch (adRegion) {
-		case 'US':
-			return 7;
-		case 'AU':
-			return 6;
-		default:
-			return 4;
-	}
 };
 
 export const realTimeConfig = (
@@ -116,10 +86,6 @@ interface AdProps extends BaseAdProps {
 	rtcConfig: string;
 }
 
-interface RegionalAdProps extends BaseAdProps {
-	config: CommercialConfig;
-}
-
 export const Ad = ({
 	isSticky = false,
 	edition,
@@ -156,48 +122,3 @@ export const Ad = ({
 		/>
 	);
 };
-
-/**
- * Ad slot component whose config differs based on region.
- * For each adRegion, create an Ad component with styling so that all but the component for the
- * user's region are hidden
- * @param RegionalAdProps
- * @returns an Ad component per region
- */
-export const RegionalAd = ({
-	edition,
-	section,
-	contentType,
-	commercialProperties,
-	config,
-}: RegionalAdProps) => (
-	<>
-		{adRegions.map((adRegion) => (
-			<ClassNames key={adRegion}>
-				{({ css, cx }) => (
-					<div
-						className={cx(
-							css`
-								${regionClasses[adRegion].styles}
-							`,
-						)}
-					>
-						<Ad
-							isSticky={false}
-							edition={edition}
-							section={section}
-							contentType={contentType}
-							commercialProperties={commercialProperties}
-							rtcConfig={realTimeConfig(
-								config.usePrebid,
-								config.usePermutive,
-								config.useAmazon,
-								getPlacementIdByAdRegion(adRegion),
-							)}
-						/>
-					</div>
-				)}
-			</ClassNames>
-		))}
-	</>
-);
