+++ dotcom-rendering/src/amp/components/Blocks.tsx
@@ -7,7 +7,7 @@ import { Elements } from '@root/src/amp/components/Elements';
 import { pillarPalette_DO_NOT_USE } from '@root/src/lib/pillars';
 import { blockLink } from '@root/src/amp/lib/block-link';
 import { findBlockAdSlots } from '@root/src/amp/lib/find-adslots';
-import { RegionalAd } from '@root/src/amp/components/Ad';
+import { RegionalAd } from '@root/src/amp/components/RegionalAd';
 
 const adStyle = css`
 	background: ${neutral[93]};
